refactor!: move components from utils dir to components dir
