+++ src/components/ActionBar/src/ActionBarLayer.vue
@@ -28,7 +28,7 @@
 <script>
 import { throttle } from 'lodash';
 import V from 'vue-v';
-import { MTransition } from '@square/maker/utils/Transition';
+import { MTransition } from '@square/maker/components/Transition';
 import { springUpBounceFn, springDownBounceFn } from '@square/maker/utils/transitions';
 import AtomicActionBar from './AtomicActionBar.vue';
