+++ src/components/Dialog/src/DialogLayer.vue
@@ -34,8 +34,8 @@
 <script>
 import Vue from 'vue';
 import PseudoWindow from 'vue-pseudo-window';
-import { MTransitionFadeIn } from '@square/maker/utils/TransitionFadeIn';
-import { MTransitionResponsive } from '@square/maker/utils/TransitionResponsive';
+import { MTransitionFadeIn } from '@square/maker/components/TransitionFadeIn';
+import { MTransitionResponsive } from '@square/maker/components/TransitionResponsive';
 import {
 	springUpFn,
 	springDownFn,
