+++ src/components/Input/src/InputFormItem.vue
@@ -25,7 +25,7 @@
 </template>
 
 <script>
-import { MBlockFormControlLayout } from '@square/maker/utils/BlockFormControlLayout';
+import { MBlockFormControlLayout } from '@square/maker/components/BlockFormControlLayout';
 import InputControl from './InputControl.vue';
 
 /**
