+++ src/components/Radio/src/RadioFormItem.vue
@@ -21,7 +21,7 @@
 </template>
 
 <script>
-import { MInlineFormControlLayout } from '@square/maker/utils/InlineFormControlLayout';
+import { MInlineFormControlLayout } from '@square/maker/components/InlineFormControlLayout';
 import RadioControl from './RadioControl.vue';
 
 /**
