+++ src/components/TransitionResize/src/TransitionResize.vue
