+++ src/components/TransitionResponsive/README.md
@@ -24,7 +24,7 @@ This is used internally by the Modal component. Resize your viewport below and a
 </template>
 
 <script>
-import { MTransitionResponsive } from '@square/maker/utils/TransitionResponsive';
+import { MTransitionResponsive } from '@square/maker/components/TransitionResponsive';
 import {
 	fadeInFn,
 	fadeOutFn,
