+++ src/components/TransitionStaggered/src/TransitionStaggered.vue
